tour handler functions moved to controller
